improve object transformation functions, add SIP bindings
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@82193 613412ce-1361-4b7a-93bf-a3c75af1bae4
