+++ pyaims/src/sip/generatedtypes.py
@@ -109,6 +109,7 @@ todo = {'system': ['Void'],
                   'aims::BucketMap<int16_t>', 'aims::BucketMap<uint16_t>',
                   'aims::BucketMap<int32_t>', 'aims::BucketMap<uint32_t>',
                   'aims::BucketMap<float>', 'aims::BucketMap<double>',
+                  'aims::Transformation3d',
                   'aims::AffineTransformation3d',
                   'carto::Site',
                   'carto::SiteIterator',
